Backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.1
uvicorn==0.27.1
python-dotenv==1.0.0
openai==1.13.3
pandas==2.2.0
pandasai==2.0.4
matplotlib==3.8.2
